removing some variables which aren't being used
line 141 and 144 create a branch by processing the image model output. However, the variables that are created after that are not being used as in line 164 when we are concatenating we are again using "image_conv_pose_xyz_flat". And also when we are finding "image_beacon_cls_fc_pose_wpqr" in line 166 we are using "image_beacon_fc_pose_xyz." Hence we can remove the creation of variables "image_beacon_fc_pose_wpqr" as well.